pyproject.toml
[tool.poetry]
name = "safebox"
version = "0.2.1"
description = "Nostr SafeBox"
authors = ["trbouma <trbouma@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
monstr = "^0.1.9"
click = "^8.1.7"
wheel = "^0.43.0"
pyyaml = "^6.0.1"
pydantic = "^2.8.2"
hotel-names = "^0.0.1"
requests = "^2.32.3"
bech32 = "^1.2.0"
mnemonic = "^0.21"
bolt11 = "^2.1.0"
qrcode = "^7.4.2"
cbor2 = "^5.6.5"
python-daemon = "^3.0.1"
bip-utils = "^2.9.3"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlmodel = "^0.0.22"
pydantic-settings = "^2.7.1"
jinja2 = "^3.1.5"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
httpx = "^0.28.1"
gunicorn = "^23.0.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
psycopg2 = "^2.9.10"
websockets = "^14.2"
hypercorn = "^0.17.3"
uvloop = "^0.21.0"
httptools = "^0.6.4"
validators = "^0.34.0"
argon2-cffi = "^23.1.0"
ipinfo = "^5.1.1"
filelock = "^3.18.0"
pyserial = "^3.5"
liboqs-python = "^0.14.1"
python-blossom = "^1.0.2"
filetype = "^1.2.0"

[tool.poetry.scripts]
safedaemon = "safebox.safedaemon:main_program"
safebox = "safebox.cli_acorn:cli"
acorn = "safebox.cli_acorn:cli"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pyinstaller = "^6.10.0"
alembic = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
